+++ src/stacks/api/report/express-app/index.ts
@@ -1,16 +1,9 @@
-import {
-  DeleteItemCommand,
-  DynamoDBClient,
-  GetItemCommand,
-  PutItemCommand,
-  QueryCommand,
-  UpdateItemCommand,
-} from '@aws-sdk/client-dynamodb';
-import { marshall, unmarshall } from '@aws-sdk/util-dynamodb';
+
 import express from 'express';
-import { AuthorizerContext, IReport, QueryStringParameters, REPORT_TABLE_NAME } from './interfaces';
+import { AuthorizerContext } from './interfaces';
 import { authorizerMiddleware } from './middlewares/authorizerMiddleware';
 import { loggerMiddleware } from './middlewares/loggerMiddleware';
+import { reportRouter } from './routes/report';
 
 declare global {
   namespace Express {
@@ -26,113 +19,8 @@ app.use(express.json());
 app.use(loggerMiddleware);
 app.use(authorizerMiddleware);
 
-const ddbClient = new DynamoDBClient();
 
-app.route('/api/report/:timestamp')
-  .get(async (req, res) => {
-    try {
-      const command = new GetItemCommand({
-        TableName: process.env[REPORT_TABLE_NAME],
-        Key: marshall({
-          ownerId: req.requestContext.authorizer.lambda.user,
-          timestamp: req.params.timestamp,
-        }),
-      });
-      const data = await ddbClient.send(command);
-      res.json(data.Item);
-    } catch (e) {
-      console.log(e);
-      res.sendStatus(500);
-    }
-  })
-  .patch(async (req, res) => {
-    try {
-      const command = new UpdateItemCommand({
-        TableName: process.env[REPORT_TABLE_NAME],
-        Key: marshall({
-          ownerId: req.requestContext.authorizer.lambda.user,
-          timestamp: req.params.timestamp,
-        }),
-        UpdateExpression: 'SET #type = :typeValue',
-        ExpressionAttributeNames: { '#type': 'type' },
-        ExpressionAttributeValues: { ':typeValue': { S: req.body.type } },
-        ReturnValues: 'ALL_NEW',
-      });
-      const data = await ddbClient.send(command);
-      res.json(data.Attributes);
-    } catch (e) {
-      console.log(e);
-      res.sendStatus(500);
-    }
-  })
-  .delete(async (req, res) => {
-    try {
-      const command = new DeleteItemCommand({
-        TableName: process.env[REPORT_TABLE_NAME],
-        Key: marshall({
-          ownerId: req.requestContext.authorizer.lambda.user,
-          timestamp: req.params.timestamp,
-        }),
-      });
-      const data = await ddbClient.send(command);
-      res.json(data);
-    } catch (e) {
-      console.log(e);
-      res.sendStatus(500);
-    }
-  });
-
-app.route('/api/report')
-  .get(async (req, res) => {
-    const limit = 2;
-    const query: QueryStringParameters = req.query;
-    try {
-      const command = new QueryCommand({
-        TableName: process.env[REPORT_TABLE_NAME],
-        KeyConditionExpression: 'ownerId = :ownerValue',
-        ExpressionAttributeValues: { ':ownerValue': { S: req.requestContext.authorizer.lambda.user } },
-        Limit: limit,
-        ExclusiveStartKey: query.LastEvaluatedKey ? marshall({
-          ownerId: req.requestContext.authorizer.lambda.user,
-          timestamp: query.LastEvaluatedKey,
-        }) : undefined,
-      });
-      const data = await ddbClient.send(command);
-      let lastEvaluatedKey;
-      if (data.LastEvaluatedKey) {
-        const lastItem = unmarshall(data.LastEvaluatedKey) as IReport;
-        lastEvaluatedKey = lastItem.timestamp;
-      }
-      let items: IReport[] = [];
-      if (data.Items) {
-        data.Items.forEach(i => items.push(unmarshall(i) as IReport));
-      }
-      res.json({
-        items: items, limit, lastEvaluatedKey,
-      });
-    } catch (e) {
-      console.log(e);
-      res.sendStatus(500);
-    }
-  })
-  .put(async (req, res) => {
-    const report: IReport = {
-      timestamp: Date.now().toString(),
-      ownerId: req.requestContext.authorizer.lambda.user,
-      type: req.body.type,
-    };
-    try {
-      const command = new PutItemCommand({
-        TableName: process.env[REPORT_TABLE_NAME],
-        Item: marshall(report),
-      });
-      await ddbClient.send(command);
-      res.json({ msg: 'OK', lastEvaluatedKey: report.timestamp });
-    } catch (e) {
-      console.error(e);
-      res.sendStatus(500);
-    }
-  });
+app.use('/api/report', reportRouter);
 
 app.route('/api/report/debug')
   .all((req, res) => {
